requirements.txt
# Core Dependencies
streamlit>=1.41.1
langchain>=0.3.13
langchain-core>=0.3.28
langchain-openai>=0.2.14
langchain-community>=0.3.11
langchain-ollama>=0.2.2
pydantic>=2.10.3
sqlalchemy>=2.0.36
tiktoken>=0.8.0

# Database
mysql-connector-python>=9.1.0
python-dotenv>=1.0.1

# Data Processing
pandas>=2.2.3
numpy>=2.2.0

# Visualization
matplotlib>=3.10.0
seaborn>=0.13.2

# Machine Learning
faiss-cpu>=1.9.0

# Document Processing
pypdf>=5.1.0

# Utilities
requests>=2.32.3
typing-extensions>=4.12.2
typing-inspect>=0.9.0
tqdm>=4.67.1

